remove redundant code for absorbing message bytes into sponge state
