+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod keccak;
+mod sponge;
 pub mod turboshake128;
 pub mod turboshake256;
